if debugging is enabled, DboSource::execute() will now add a tag (comment) to the query sent to the database.
The tag is the part of the stacktrace that matches 'APP/models'. Usually you will get one or two entries on a single line.
This is a useful complement to the already existing logging of queries because now you can more easily where queries are coming from,
when checking your database server logs, slow query log, current process list, when investigating locks, etc.