integrate basic http authentication into Auth component
If http authentication is enabled (use case: a portal where you login once, to access multiple applications),
it is convenient to integrate this into Cake.

In the Auth component startup function:
if:
	* the user is not logged in
	* http auth integration is enabled (disabled by default)
	* $_SERVER['PHP_AUTH_USER'] is set
	* the user is found in the database

We will automatically log in this user.
No need to check passwords, because we trust the http authentication.